Update CRI-O commit id to fix failing CRI-O jobs
This should potentiall fix failing CRI-O jobs caused by
https://github.com/kubernetes/test-infra/pull/31501

That PR uses CRI-O commit that doesn't have a correspondent tarball
at https://storage.googleapis.com/cri-o/artifacts
Ignition script fails to download the tarbal and fails, which
perheps results in a boot failure for the test instance.

This commit changes the next commit id to the one that has
a tarball. However, this commit doesn't point to the CRI-O
release tag, unlike the one that was used in #31501.