add metrics for job config ConfigMap(s)
Kubernetes limits ConfigMap objects to 1MB. We store job configs in
ConfigMap objects, but don't track the sizes of these ConfigMaps.

There is some prior art in 589f8431d9 (Merge pull request #13376 from
stevekuznetsov/skuznets/configmap-sizes, 2019-07-10) but that metric
only works for plugins (in "hook"). For Prow instances using Gerrit,
those metrics won't be available.

Add metrics in Sinker specifically for the job configurations, whose
ConfigMap(s) are mounted into the filesystem with --job-config-path. We
have to check if the path mounts a single ConfigMap, or if the path has
subdirectories (each of which are ConfigMaps) when the job config is
partitioned into multiple ConfigMaps, per 60dadd267c (Merge pull
request #28835 from cjwagner/partitioned-configmaps, 2023-02-23). We use
the "name" label in the metric to record the name of the ConfigMap
directory we got the size for.

The reason we read from disk (instead of getting the ConfigMap back
directly from the Kubernetes API server) is because we should try to
avoid hitting the API server when possible (Prow uses Prow Job
custom resources, which can number in the tens of thousands in prod, and
so the API server can be quite busy).