Apply blobless_fetch to extra_refs
Set BloblessFetch in ExtraRefs based on the DecorationConfig.
This ensures that periodic jobs (which only specify ExtraRefs) and
any presubmit, etc. jobs that happen to specify extra_refs use
blobless clones consistently.

Setting BloblessFetch in ExtraRefs will not override the default
config, due to its boolean type. In order to override the default for
a job, BloblessConfig must be set in the DecorationConfig instead.