bump Prow base images
Follow-up of https://github.com/kubernetes/test-infra/pull/31745. This
picks up Alpine Edge for external dependencies outside of our Go-based
binaries. For example, for the `git` base image, this brings in libcurl
8.5.0 which addresses several CVEs for curl's dependencies.

In the future we can move away from alpine (to distroless) for those
images that don't depend on any additional Alpine packages (these are
all the ones that just use plain "alpine" as the base image).