chore: revert prow cluster move for capg
This change reverts #32426 in which the remaining prow jobs where moved
to the EKS cluster. This is because some resource are not currently
available via boskos as seen by this error:

```
boskos: checkout_account: Got invalid response while checking for 'free'
: 404: Not Found, b'Acquire failed: resource type "gce-project" does
not exist\n
```

This is temporary whilst we investigate.

Signed-off-by: Richard Case <richard.case@suse.com>